Import GPX tracks

Import GPX tracks from commercial GPS equipment as a data.table ready for velocity function estimation. Note that the output coordinates must still be converted to a conformal projection in meters if they are to be used in functions other than getVelocity, dtVelocity, or downsampleXYZ.

importGPX(tracks, verbose = FALSE)

Arguments

tracks

A character string or vector with filepaths pointing to the location of the GPX files

verbose

Should a progress bar be printed? Default is verbose = FALSE, recommended particularly when used inside lapply

Value

A data.table with five or six columns:

(1) $TrackID The name of the input GPX track

(2) $PID The order that point appeared in the GPX file

(3) $t The timestamp

(4) $long The longitude in decimal degrees

(5) $lat The latitude in decimal degrees

(6) $z The elevation (if present)

Examples

# Get a list of GPX tracks in a directory
gpx <-  list.files(pattern = ".gpx$")

# Convert to data.table
gpx <- importGPX(gpx)